requirements2.txt
certifi==2023.7.22
chardet==4.0.0
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.2.0
cycler==0.10.0
defusedxml==0.7.1
filelock==3.13.1
fonttools==4.48.1
fsspec==2024.2.0
idna==2.10
Jinja2==3.1.3
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.2
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
opencv-python==4.9.0.80
packaging==23.2
pandas==2.2.0
pillow==10.2.0
psutil==5.9.8
py-cpuinfo==9.0.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-magic==0.4.27
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
requests-toolbelt==1.0.0
roboflow==1.1.19
scipy==1.12.0
seaborn==0.13.2
six==1.16.0
supervision==0.17.1
sympy==1.12
thop==0.1.1.post2209072238
torch==2.2.1+cu118
torchaudio==2.2.1+cu118
torchvision==0.17.1+cu118
tqdm==4.66.2
typing_extensions==4.9.0
tzdata==2024.1
ultralytics==8.0.227
urllib3==2.2.0
